added the automatic firing of multiple dma transactions if the transfer size is over 65535 bytes. This should work for poling or asynchronous dma transfers. * addDescriptor and changeDescriptor in Adafruit_ZeroDMA function signature changed. Count parameter changed to be a uint16_t instead of uint32_t. Reflects the limitation that desc->BTCNT.reg is a uint16_t and cannot actually accept a uint32_t value. * Not tested on hardware yet, pushed so it can be read by others. Will test on hardware soon. Need to make sure line 312 and 313 pointer assignment math in checkDmaComplete actually works properly.

+++ libraries/Adafruit_ZeroDMA/Adafruit_ZeroDMA.cpp
@@ -449,7 +449,7 @@ bool Adafruit_ZeroDMA::isActive(void) const {
 DmacDescriptor *Adafruit_ZeroDMA::addDescriptor(
   void           *src,
   void           *dst,
-  uint32_t        count,
+  uint16_t        count,
   dma_beat_size   size,
   bool            srcInc,
   bool            dstInc,
@@ -537,7 +537,7 @@ DmacDescriptor *Adafruit_ZeroDMA::addDescriptor(
 // etc.) are unchanged.  Mostly for changing the data being pushed to a
 // peripheral (DAC, SPI, whatev.)
 void Adafruit_ZeroDMA::changeDescriptor(DmacDescriptor *desc,
-  void *src, void *dst, uint32_t count) {
+  void *src, void *dst, uint16_t count) {
 
     uint8_t bytesPerBeat; // Beat transfer size IN BYTES
     switch(desc->BTCTRL.bit.BEATSIZE) {
