+++ libraries/Adafruit_ZeroDMA/Adafruit_ZeroDMA.h
@@ -41,13 +41,13 @@ class Adafruit_ZeroDMA {
   uint8_t         getChannel(void) const { return channel; }
 
   // DMA descriptor functions
-  DmacDescriptor *addDescriptor(void *src, void *dst, uint32_t count = 0,
+  DmacDescriptor *addDescriptor(void *src, void *dst, uint16_t count = 0,
                     dma_beat_size size = DMA_BEAT_SIZE_BYTE,
                     bool srcInc = true, bool dstInc = true, 
                     uint32_t stepSize = DMA_ADDRESS_INCREMENT_STEP_SIZE_1, 
                     bool stepSel = DMA_STEPSEL_DST);
   void            changeDescriptor(DmacDescriptor *d, void *src = NULL,
-                    void *dst = NULL, uint32_t count = 0);
+                    void *dst = NULL, uint16_t count = 0);
   bool            isActive(void) const;
 
   void            _IRQhandler(uint8_t flags); // DO NOT TOUCH
